% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliaryArribaFunctions.R
\name{.ReadArribaFusionTable}
\alias{.ReadArribaFusionTable}
\title{.ReadArribaFustionTable
Read and format arriba output fusion table}
\usage{
.ReadArribaFusionTable(fusionsFile)
}
\arguments{
\item{fusionFile}{string with the full path of the Fussion.tsv table}
}
\description{
.ReadArribaFustionTable
Read and format arriba output fusion table
}
