% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliaryArribaFunctions.R
\name{.ReadCytobandFile}
\alias{.ReadCytobandFile}
\title{.ReadCytobandFile
Internal function. Read cytoband information}
\usage{
.ReadCytobandFile()
}
\arguments{
\item{fusionFile}{string with the full path of the Fussion.tsv table}
}
\description{
.ReadCytobandFile
Internal function. Read cytoband information
}
