% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliaryArribaFunctions.R
\name{.StatsSTARfile}
\alias{.StatsSTARfile}
\title{.StatsSTARfile
reads and format the Log-final.out alignement statistics file
This is an internal function, it should not be directly used}
\usage{
.StatsSTARfile(bamFile)
}
\arguments{
\item{bamFile}{the prefix_Aligned.out.bam or prefix_SortByCoordinates.bam file}
}
\value{
a data frame with the Log content
}
\description{
.StatsSTARfile
reads and format the Log-final.out alignement statistics file
This is an internal function, it should not be directly used
}
