% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliaryArribaPlotFunctions.R
\name{fusionPlot}
\alias{fusionPlot}
\title{fusionPlot 
This function provides gene fusion visualization}
\usage{
fusionPlot(sbjBamFile, ...)
}
\arguments{
\item{sbjBamFile}{string with full path of the processed and sorted bam file processed by runSTAR and runArriba}
}
\description{
fusionPlot 
This function provides gene fusion visualization
}
\seealso{
runArriba
}
