% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runners2.R
\name{GetArribaRTest}
\alias{GetArribaRTest}
\title{GetArribaRTest
This function will retrieve the test files coming with the arriba software}
\usage{
test.files <- GetArribaRTest()
}
\value{
an string with the full path of the firs read file of the test
}
\description{
GetArribaRTest
This function will retrieve the test files coming with the arriba software
}
