% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneExpression.R
\name{GetCounts}
\alias{GetCounts}
\title{GetCounts
it estimate the gene counts wrapping the Rsubread::featureCounts function}
\usage{
ft <- GetCounts("my.aligned_bam", feature = "exon")
}
\arguments{
\item{sbjBAM}{the BAM file name}

\item{feature}{string indicating if exons or genes are returned}

\item{nthreads}{integer, total of treads to use, default parallel::detectCores() -1}
}
\description{
GetCounts
it estimate the gene counts wrapping the Rsubread::featureCounts function
}
