% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneExpression.R
\name{GetImmuneContent}
\alias{GetImmuneContent}
\title{GetImmuneContent}
\usage{
GetImmuneContent(ftc, filter = 0, plot = TRUE)
}
\arguments{
\item{ftc}{object returned by \code{\link{GetCounts}} The ftc should be an object returned by  \code{\link{GetCounts}} called with the parameter feature = "gene"
this parameter is saved in ftc$targets[2] position and this function will check this, otherwise it trough an error}

\item{filter}{(integer) minimal count level to consider (default 0L)}

\item{plot}{(logical) if TRUE (default) the subject proportion plot, no plot otherwise}
}
\value{
a MIXTURE object (see \code{\link[MIXTURE]{MIXTURE}})
}
\description{
Estimate the Tissue Immune Microenvironment content based on LM22 signature. It is based on TPMs calculated from the gene counts. duplicated genes will be removed
If MIXTURE package is not already installed, it will be installed
}
\examples{
\dontrun{
test.subject <- GetArribaRTest()
bam.subject <- RunSTAR(test.subject)
ft <- GetCounts(bam.subject)
ft.imm <- GetImmuneContent(ft)
}
}
