% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runners.R
\name{RunSortIndexBam}
\alias{RunSortIndexBam}
\title{runSortIndexBam
This function will built the sorted bam file and its index.}
\usage{
RunSortIndexBam(sbjBamFile, remove = T)
}
\arguments{
\item{sbjBamFile}{string full path of aligned bam file see \code{\link[runSTAR]{runSTAR}}}

\item{remove}{boolean (default  TRUE), if unsorted bam file should be removed or not.}
}
\description{
runSortIndexBam
This function will built the sorted bam file and its index.
}
\seealso{
\code{\link[Rsamtools]{sortBam}}
}
