% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runners.R
\name{runSTAR}
\alias{runSTAR}
\title{runSTAR
This function will run STAR software}
\usage{
runSTAR(sbjFile)
}
\arguments{
\item{sbjFile}{string with the full path and name of the xxxx_1_fastq or gz sequence file. This
function only support paired data}
}
\description{
runSTAR
This function will run STAR software
}
\examples{
\dontrun{
test.subject <- getArribaRTest()
bam.subject <- runSTAR(test.subject)
}
}
\author{
Elmer A. Fernández
}
