% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliaryArribaFunctions.R
\name{SearchPanel}
\alias{SearchPanel}
\title{SearchPanel
This function recover those fusions that were discarded by Arriba, 
but they have at split_read1+spleat_read2+discordant_mates > 0. Low confident fusions}
\usage{
SearchPanel(genePanel, fileFusions, save = T)
}
\arguments{
\item{genePanel}{a data frame with the target genes}

\item{fileFusions}{the tsv discarded file fusions path, it should end in _Fusions_discarded.tsv}

\item{save}{(logical, default TRUE), if the results would be saved as excel file}
}
\description{
SearchPanel
This function recover those fusions that were discarded by Arriba, 
but they have at split_read1+spleat_read2+discordant_mates > 0. Low confident fusions
}
