% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliaryArribaFunctions.R
\name{.ReadCytobandFile}
\alias{.ReadCytobandFile}
\title{.ReadCytobandFile
Internal function. Read cytoband information}
\usage{
.ReadCytobandFile(species, version)
}
\arguments{
\item{species}{see \code{\link[GenomeDB]{GetGenomes}}}

\item{version}{see \code{\link[GenomeDB]{GetGenomes}}}
}
\description{
.ReadCytobandFile
Internal function. Read cytoband information
}
