% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliaryArribaPlotFunctions.R
\name{FusionPlot}
\alias{FusionPlot}
\title{FusionPlot 
This function provides gene fusion visualization}
\usage{
FusionPlot(sbjBamFile,gene1,gene2)
}
\arguments{
\item{sbjBamFile}{string with full path of the processed and sorted bam file processed by runSTAR and runArriba}

\item{gene1}{string with geneID as returned by \code{\link{RunArriba}} 
_Fusion.xlsx table. Default missing (will plot all the fusions), otherwise it will only plot all the fusions where this gene is present}

\item{gene2}{idem as gene1. If both gene1 and gene2 are set, only the fusions including those genes will be plotted}
}
\description{
FusionPlot 
This function provides gene fusion visualization
}
\examples{
\dontrun{
test.subject <- GetArribaRTest()
bam.subject <- RunSTAR(test.subject)
Fusions <- RunArriba(bam.subject)
bam.sorted.subject <- RunSortIndexBam(bam.subject)
FusionPlot(bam.sorted.subject)
}

}
\seealso{
\code{\link{runArriba}}
}
