% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Installers.R
\name{InstallArribaR}
\alias{InstallArribaR}
\title{InstallArribaR
This function will install the arri software and all its dependencies like the roght STAR version
It currently install version 2.1.0 of arriba and 2.7.6a of STAR, the GRCh37 genome assembly and GENCODE19.gtf annotation
future installer will be more felxible}
\usage{
InstallArribaR(extDir)
}
\arguments{
\item{extDir}{character string of the instalation path. It should be write enable and hold enouph space
if missing, it will create on the current directory the folder "Software" and everything will be installed there
if given, it will create the folder givenpath/Software}
}
\description{
InstallArribaR
This function will install the arri software and all its dependencies like the roght STAR version
It currently install version 2.1.0 of arriba and 2.7.6a of STAR, the GRCh37 genome assembly and GENCODE19.gtf annotation
future installer will be more felxible
}
