% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runners.R
\name{runArriba}
\alias{runArriba}
\title{runArriba
this function will run the gene fusion detection ARRIBA software}
\usage{
runArriba(sbjBamFile)
}
\arguments{
\item{sbjBamFile}{string with the full file name of the subject BAM file}
}
\value{
a data frame with the identified fusions
\url{https://arriba.readthedocs.io/en/latest/output-files/}{see details in Arriba output files}
}
\description{
runArriba
this function will run the gene fusion detection ARRIBA software
}
\examples{
\dontrun{
test.subject <- getArribaRTest()
bam.subject <- runSTAR(test.subject)
Fusions <- runArriba(bam.subject)
View(Fusions)
}
}
\references{
\url{https://genome.cshlp.org/content/early/2021/02/11/gr.257246.119}{Uhrig et al.}
}
\seealso{
\code{\link{runSTAR()}}
}
